Create generate forecast file
Relates #27
