+++ utils/submitEventListener.js
@@ -3,9 +3,9 @@ import {
   hideBufferingOverlay,
 } from "./helpers/bufferingOverlay.js";
 import { createElement, appendElements } from "./helpers/handleElements.js";
-import { getFormattedWeekdate } from "./helpers/getFormattedWeekDate.js";
 import { getMessage } from "./api/getMessage.js";
 import { handleApiRequests } from "./handleApiRequests.js";
+import { generateForecast } from "./generateForecast.js";
 
 export const submitEventListener = () => {
   const form = document.querySelector("form");
@@ -18,7 +18,7 @@ export const submitEventListener = () => {
   const forecastMessage = document.querySelector(".forecast-message");
   const invalidMessage = document.querySelector(".invalid-message");
   const mapDiv = document.querySelector("#location");
-  const weekdays = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
+
   let image = null;
 
   form.addEventListener("submit", async (event) => {
@@ -133,38 +133,7 @@ export const submitEventListener = () => {
         appendElements(currSunrise, [sunrise_icon, sunrise]);
         appendElements(currSunset, [sunset_icon, sunset]);
 
-        for (let i = 1; i < 6; i++) {
-          const weekday = new Date(date_arr[i - 1]).getDay();
-          const weekdate = getFormattedWeekdate(weekday);
-          const forecast_icon = createElement("img", {
-            src: `./utils/icons/${fore_code[i - 1]}.svg`,
-            className: "icon",
-            alt: "forecast icon",
-          });
-          const forecast_max = createElement("h4");
-          forecast_max.textContent = `${fore_max[i - 1]}°`;
-
-          const forecast_min = createElement("h4");
-          forecast_min.textContent = `${fore_min[i - 1]}°`;
-
-          const day = createElement("div");
-
-          const forecast_date = createElement("h4");
-          forecast_date.textContent = weekdate;
-
-          const forecast_day = createElement("h4");
-          forecast_day.textContent = weekdays[weekday];
-
-          appendElements(day, [
-            forecast_date,
-            forecast_day,
-            forecast_icon,
-            forecast_max,
-            forecast_min,
-          ]);
-          document.querySelector(`#day${i}`).innerHTML = "";
-          document.querySelector(`#day${i}`).appendChild(day);
-        }
+        generateForecast(date_arr, fore_code, fore_max, fore_min);
         hideBufferingOverlay();
       } else {
         throw new Error(response.status);
